refactor: use u64 for countable types (event_count, size_bytes, version_count)
These values come from COUNT(*) or file sizes and are never negative.
Changed from i64 to u64 in storage types and DTOs. DB binding casts as
i64 (SQLite INTEGER limitation). Renamed U32OutOfRange to IntegerCast
with a target field for reuse.

+++ aperture-artifacts/src/artifacts.rs
@@ -503,7 +503,7 @@ fn build_artifact(
         digest: digest.to_string(),
         media_type: Some(media_type.to_string()),
         version,
-        size_bytes: size as i64,
+        size_bytes: size,
         downloaded_at: at,
         verified_at: Some(at),
     }
